requirements.txt
numpy
PyYAML
Pillow
opencv-python
requests
cairosvg
scipy
scikit-image
